🚚 Rename: LinkCardCollection을 분리 및 폴더 구조 시인성 강화
