+++ src/components/link-card/LinkCard.style.tsx
@@ -1,20 +1,5 @@
 import styled from "styled-components";
 
-export const CardGridLayout = styled.section`
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  justify-items: center;
-  gap: 2.5rem 2rem;
-
-  @media (max-width: 1124px) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media (max-width: 767px) {
-    grid-template-columns: 1fr;
-  }
-`;
-
 export const CardBoxOriginPosition = styled.div`
   padding: 0;
   border: none;
@@ -81,7 +66,7 @@ export const CardImageDiv = styled.div<{ $image?: string }>`
   background-image: url(${({ $image }) =>
     $image
       ? `${$image}); background-size: cover`
-      : `/assets/images/svg/no-img-card.svg)`};
+      : `/assets/images/svg/no-img-card.svg)`};;
 `;
 
 export const TextSection = styled.div`
