+++ src/components/link-card/link-card-layout/LinkCardListLayout.tsx
@@ -1,0 +1,38 @@
+import * as S from "./LinkCardListLayout.style";
+
+import UserLinkDataType from "@/src/types/UserLinkDataType";
+import { LinkCardFunctionObjectType } from "@/src/types/ModalFunctionDataTypes";
+
+import LinkCard from "../LinkCard";
+
+interface LinkCardListLayoutPropType {
+  items: UserLinkDataType[];
+  favorite?: boolean;
+  kebab?: LinkCardFunctionObjectType[];
+}
+
+/**
+ *
+ * @param { FolderCardData } items 카드에 대한 전반적인 내용이 담긴 요소입니다.
+ * @param { boolean } favorite 현재 카드의 즐겨찾기 여부를 표시하는 요소입니다.
+ * @param { FolderKebabActionArray } kebab 케밥 메뉴에 대한 동작들이 포함된 컴포넌트입니다.
+ * @returns 링크 카드의 그리드 배열을 구현하는 단순 ui 배치형 컴포넌트입니다.
+ */
+export default function LinkCardListLayout({
+  items,
+  favorite = false,
+  kebab,
+}: LinkCardListLayoutPropType) {
+  return (
+    <S.CardGridLayout>
+      {items.map((item) => (
+        <LinkCard
+          key={item.id}
+          contents={item}
+          favorite={favorite}
+          kebab={kebab}
+        />
+      ))}
+    </S.CardGridLayout>
+  );
+}
